Removed axios - will just use native fetch API refactoring Graphvisualisation
